product_enhancer_agent/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
requests
python-dotenv
